HAL-1565: Add [] as allowed characters in properties form item
