+++ ballroom/src/main/java/org/jboss/hal/ballroom/form/PropertiesItem.java
@@ -61,7 +61,7 @@ public class PropertiesItem extends TagsItem<Map<String, String>> {
 
     @Override
     public String allowedCharacters() {
-        return "- . : @ ; = ? ! # $ % &";
+        return "- . : @ ; = ? ! # $ % & [ ]";
     }
 
     @Override
@@ -74,10 +74,11 @@ public class PropertiesItem extends TagsItem<Map<String, String>> {
         }
     }
 
-    private static class MapMapping implements TagsMapping<Map<String, String>> {
+
+    static class MapMapping implements TagsMapping<Map<String, String>> {
 
         private static final RegExp REGEX = RegExp.compile(
-                "^([\\w\\-\\.\\/]+)=([\\w\\-\\.\\/:\\@\\;\\=\\?\\!\\#\\$\\%\\&]+)$"); //NON-NLS
+                "^([\\w\\-\\.\\/]+)=([\\w\\-\\.\\/:\\@\\;\\=\\?\\!\\#\\$\\%\\&\\[\\]\\,]+)$"); //NON-NLS
         private static final String EQ = "=";
 
         @Override
