+++ meta/src/test/java/org/jboss/hal/meta/security/ConstraintsTest.java
@@ -16,74 +16,125 @@
 package org.jboss.hal.meta.security;
 
 import org.jboss.hal.meta.AddressTemplate;
-import org.junit.Before;
 import org.junit.Test;
 
-import static org.jboss.hal.dmr.ModelDescriptionConstants.ADD;
-import static org.jboss.hal.dmr.ModelDescriptionConstants.ENABLED;
 import static org.jboss.hal.meta.security.Constraints.Operator.AND;
 import static org.jboss.hal.meta.security.Constraints.Operator.OR;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-@SuppressWarnings({"HardCodedStringLiteral", "DuplicateStringLiteralInspection"})
 public class ConstraintsTest {
 
-    private AddressTemplate template;
+    private static final Constraint ENGAGE = Constraint.executable(AddressTemplate.of("j/l/p"), "engage");
+    private static final Constraint NCC = Constraint.writable(AddressTemplate.of("da/ta"), "NCC-1701-D");
+    private static final String ENGAGE_DATA = "executable(j/l/p:engage)";
+    private static final String NCC_DATA = "writable(da/ta@NCC-1701-D)";
+    private static final String AND_DATA = ENGAGE_DATA + "&" + NCC_DATA;
+    private static final String OR_DATA = ENGAGE_DATA + "|" + NCC_DATA;
 
-    @Before
-    public void setUp() throws Exception {
-        template = AddressTemplate.of("{selected.profile}/subsystem=datasources/data-source=*");
+    @Test
+    public void empty() {
+        Constraints empty = Constraints.empty();
+        assertEmpty(empty);
     }
 
     @Test
-    public void parseNull() throws Exception {
-        assertTrue(Constraints.parse(null).isEmpty());
+    public void single() {
+        Constraints single = Constraints.single(ENGAGE);
+        assertSingle(single);
     }
 
     @Test
-    public void parseEmpty() throws Exception {
-        assertTrue(Constraints.parse("").isEmpty());
+    public void and() {
+        Constraints and = Constraints.and(ENGAGE, NCC);
+        assertAnd(and);
     }
 
     @Test
-    public void parseBlank() throws Exception {
-        assertTrue(Constraints.parse("   ").isEmpty());
+    public void or() {
+        Constraints or = Constraints.or(ENGAGE, NCC);
+        assertOr(or);
     }
 
     @Test
-    public void parseInvalid() throws Exception {
-        assertTrue(Constraints.parse("foo").isEmpty());
+    public void parseSingle() {
+        Constraints single = Constraints.parse(ENGAGE_DATA);
+        assertSingle(single);
     }
 
     @Test
-    public void parseOne() throws Exception {
-        String input = "executable({selected.profile}/subsystem=datasources/data-source=*:add)";
-        Constraints constraints = Constraints.parse(input);
-
-        assertEquals(1, constraints.size());
-        assertTrue(constraints.contains(Constraint.executable(template, ADD)));
+    public void parseAnd() {
+        Constraints and = Constraints.parse(AND_DATA);
+        assertAnd(and);
     }
 
     @Test
-    public void parseAnd() throws Exception {
-        String input = "executable({selected.profile}/subsystem=datasources/data-source=*:add)&writable({selected.profile}/subsystem=datasources/data-source=*@enabled)";
-        Constraints constraints = Constraints.parse(input);
-
-        assertEquals(2, constraints.size());
-        assertEquals(AND, constraints.getOperator());
-        assertTrue(constraints.contains(Constraint.executable(template, ADD)));
-        assertTrue(constraints.contains(Constraint.writable(template, ENABLED)));
+    public void parseOr() {
+        Constraints or = Constraints.parse(OR_DATA);
+        assertOr(or);
     }
 
     @Test
-    public void parseOr() throws Exception {
-        String input = "executable({selected.profile}/subsystem=datasources/data-source=*:add)|writable({selected.profile}/subsystem=datasources/data-source=*@enabled)";
-        Constraints constraints = Constraints.parse(input);
+    public void parseNull() {
+        Constraints constraints = Constraints.parse(null);
+        assertEmpty(constraints);
+    }
 
-        assertEquals(2, constraints.size());
-        assertEquals(OR, constraints.getOperator());
-        assertTrue(constraints.contains(Constraint.executable(template, ADD)));
-        assertTrue(constraints.contains(Constraint.writable(template, ENABLED)));
+    @Test
+    public void parseEmpty() {
+        Constraints constraints = Constraints.parse("");
+        assertEmpty(constraints);
+    }
+
+    @Test
+    public void parseBlank() {
+        Constraints constraints = Constraints.parse("   ");
+        assertEmpty(constraints);
+    }
+
+    @Test
+    public void parseIllegal() {
+        Constraints constraints = Constraints.parse("no constraints");
+        assertEmpty(constraints);
+    }
+
+
+    // ------------------------------------------------------ helper methods
+
+    private void assertEmpty(Constraints empty) {
+        assertTrue(empty.isEmpty());
+        assertTrue(empty.getConstraints().isEmpty());
+        assertEquals(AND, empty.getOperator());
+        assertEquals(0, empty.size());
+        assertEquals("", empty.data());
+        assertEquals("", empty.toString());
+    }
+
+    private void assertSingle(Constraints single) {
+        assertFalse(single.isEmpty());
+        assertFalse(single.getConstraints().isEmpty());
+        assertEquals(AND, single.getOperator());
+        assertEquals(1, single.size());
+        assertEquals(ENGAGE.data(), single.data());
+        assertEquals(ENGAGE.toString(), single.toString());
+    }
+
+    private void assertAnd(Constraints and) {
+        assertFalse(and.isEmpty());
+        assertFalse(and.getConstraints().isEmpty());
+        assertEquals(AND, and.getOperator());
+        assertEquals(2, and.size());
+        assertEquals(AND_DATA, and.data());
+        assertEquals(AND_DATA, and.toString());
+    }
+
+    private void assertOr(Constraints or) {
+        assertFalse(or.isEmpty());
+        assertFalse(or.getConstraints().isEmpty());
+        assertEquals(OR, or.getOperator());
+        assertEquals(2, or.size());
+        assertEquals(OR_DATA, or.data());
+        assertEquals(OR_DATA, or.toString());
     }
 }
